// week-5/in-n-out-booksp3/src/app/book.interface.ts
/*
 Title: book.interface.ts
 Author: Professor Krasso
 Modified By: April Yang
 Date: 11/20/2022
 Description: Assignment 5.4 - Dialogs
Resources: Material Design https://material.io
*/

export interface IBook{
  isbn: string;
  title: string;
  description: string,
  numOfPages: number;
  authors: Array<string>;
}